refactor(server): extract mode-aware plugin-dispatch seam for dev+preview reuse
Move the dev router's private plugin-dispatch / security helpers out of
routes.rs into plugin_middleware.rs so the wave-3 preview server can consume
the same code instead of a drifting reimplementation. Pure, behaviour-
preserving refactor — no functional change.

Extracted into plugin_middleware.rs (all mode-aware):
- dispatch_plugin + PluginDispatchAttempt (plugin request build/dispatch,
  base64 body decode, header passthrough, no-store cache posture)
- origin_gate (#931 Origin rejection for non-GET requests; takes a
  HostValidation which already carries the ServerMode, so no separate
  mode param needed)
- plugin_error_response (#926 mode-gated verbosity: verbose in dev,
  generic in preview/embed)
- PLUGIN_BODY_LIMIT_BYTES (2 MiB) + body_limit_layer()

routes.rs now delegates: origin_rejection is a thin AppState wrapper over
origin_gate; build_router uses body_limit_layer(); the dev plugin leg in
serve_page calls the shared dispatch_plugin. Dev behaviour is byte-for-byte
identical (verified by the untouched zfb-server test suite).

Refs #1544
